pyproject.toml
[project]
name = "dsas"
version = "0.3.0"
description = "A Multi-agent deep research system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "openai>=1.40.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.1",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "tenacity>=8.4.2",
  "openai-agents>=0.2.10",
  "trafilatura>=1.9.0",
  "pypdf>=4.3.1",
  "beautifulsoup4>=4.13.5",
  "lxml>=5.4.0",
]
[tool.uv]
